requirements.txt
django-registration
django-registration-defaults
pycryptodomex
demjson
django
rpyc
pymongo
requests
